Renamed model namespace to models.
For consistency with views and (planned) events.
